Rename viewer/model.py to viewer_model.py
