+++ napari/components/_viewer/viewer_model.py
